Add ping to tab title when there are events
